Fixed discrepancy in right/left movement behaviour
'Acceleration' is applied every game cycle, irrespective of whether objects are actually in motion. If the playerXV variable is not set to zero initially or after ceasing movement, then this variable rapidly increases to the maximum allowed speed. The original code would set playerXV to 1.0 (==maxSpeed). The result is that movement in the positive X direction would instantaneously commence at maximum speed.