Fixed unexpected response to mouse clicks
As the square falls, following a release of the mouse, additional clicks will result in the square teleporting to the bottom. This may be intended but I found it undesirable. In this version, additional clicks on the falling square will 'catch' it (as before) but clicks elsewhere will have no effect.